# docs/swagger.yaml
basePath: /
definitions:
...
  requests.PrintRequestPayload:
    properties:
      approver_id:
        type: integer
      bw_copies:
        type: integer
      color_copies:
        type: integer
      description:
        type: string
      requested_at:
        type: string
      requester_id:
        type: integer
    type: object
  requests.PrintRequestUpdatePayload:
    properties:
      approver_id:
        type: integer
      bw_copies:
        type: integer
      color_copies:
        type: integer
      description:
        type: string
      requested_at:
        type: string
      requester_id:
        type: integer
    type: object
  requests.PublisherPayload:
    properties:
      name:
        type: string
    type: object
  requests.PublisherUpdatePayload:
    properties:
      name:
        type: string
    type: object
...
  resources.ApproverResource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  resources.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
  resources.PrintRequestCollection:
    properties:
      data:
        items:
          $ref: '#/definitions/resources.PrintRequestResource'
        type: array
      meta:
        $ref: '#/definitions/resources.PaginationMeta'
    type: object
  resources.PrintRequestResource:
    properties:
      approver:
        $ref: '#/definitions/resources.ApproverResource'
      approver_id:
        type: integer
      bw_copies:
        type: integer
      color_copies:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      requested_at:
        type: string
      requester:
        $ref: '#/definitions/resources.RequesterResource'
      requester_id:
        type: integer
      updated_at:
        type: string
    type: object
  resources.PublisherCollection:
    properties:
      data:
        items:
          $ref: '#/definitions/resources.PublisherResource'
        type: array
      meta:
        $ref: '#/definitions/resources.PaginationMeta'
    type: object
  resources.PublisherResource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  resources.RequesterResource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.ErrorBody:
    properties:
      details: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  title: Safa API
  version: "1.0"
paths:
...
  /print-requests:
    get:
      description: Get a paginated list of print requests with optional filters
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 15)
        in: query
        name: per_page
        type: integer
      - description: Comma-separated requester names
        in: query
        name: requester_names
        type: string
      - description: Comma-separated approver names
        in: query
        name: approver_names
        type: string
      - description: Minimum color copies
        in: query
        name: color_copies_min
        type: integer
      - description: Maximum color copies
        in: query
        name: color_copies_max
        type: integer
      - description: Minimum BW copies
        in: query
        name: bw_copies_min
        type: integer
      - description: Maximum BW copies
        in: query
        name: bw_copies_max
        type: integer
      - description: Start date (RFC3339)
        in: query
        name: requested_at_from
        type: string
      - description: End date (RFC3339)
        in: query
        name: requested_at_to
        type: string
      - description: Sort column
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.PrintRequestCollection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: List all PrintRequest records
      tags:
      - PrintRequest
    post:
      consumes:
      - application/json
      parameters:
      - description: PrintRequest payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.PrintRequestPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.PrintRequestResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a PrintRequest
      tags:
      - PrintRequest
  /print-requests/{id}:
    delete:
      parameters:
      - description: PrintRequest ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete a PrintRequest
      tags:
      - PrintRequest
    get:
      parameters:
      - description: PrintRequest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.PrintRequestResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get a PrintRequest by ID
      tags:
      - PrintRequest
    put:
      consumes:
      - application/json
      parameters:
      - description: PrintRequest ID
        in: path
        name: id
        required: true
        type: integer
      - description: PrintRequest update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.PrintRequestUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.PrintRequestResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update a PrintRequest
      tags:
      - PrintRequest
  /publishers:
    get:
      description: Get a paginated list of Publisher records
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 15)
        in: query
        name: per_page
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.PublisherCollection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: List all Publisher records
      tags:
      - Publisher
    post:
      consumes:
      - application/json
      parameters:
      - description: Publisher payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.PublisherPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.PublisherResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a Publisher
      tags:
      - Publisher
  /publishers/{id}:
    delete:
      parameters:
      - description: Publisher ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete a Publisher
      tags:
      - Publisher
    get:
      parameters:
      - description: Publisher ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.PublisherResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get a Publisher by ID
      tags:
      - Publisher
    put:
      consumes:
      - application/json
      parameters:
      - description: Publisher ID
        in: path
        name: id
        required: true
        type: integer
      - description: Publisher update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.PublisherUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.PublisherResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update a Publisher
      tags:
      - Publisher
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"